Фиксы и чистка кода

+++ swapget/db_to_graph.py
@@ -11,7 +11,8 @@ def create_graph_from_db(engine, block_number: str, tokens_table: dict) -> nx.Di
         for row in result:
             token0_index = tokens_table[row.token0_address]
             token1_index = tokens_table[row.token1_address]
-            weight = [int(row.token0_reserve), int(row.token1_reserve)]
+            weight = [float(row.token0_reserve), float(row.token1_reserve)]
+
             graph.add_edge(token0_index, token1_index, weight=weight)
 
     return graph
